api/pyproject.toml
[project]
name = "graphsuite-api"
version = "0.1.0"
description = "GraphSuite HTTP API (Connect, Governance, Ontology gateway)."
readme = "README.ja.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "GraphSuite", email = "contact@example.com"}
]
keywords = ["graphsuite", "fastapi", "governance", "ontology", "connect", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "PyJWT[crypto]>=2.8.0",
    "boto3>=1.34.0",
    "python-multipart>=0.0.6",
    "ascii-colors>=0.4.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
]

production = [
    "gunicorn>=21.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/graphsuite"
Repository = "https://github.com/your-org/graphsuite.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
include = [
    "app/**/*.py",
    "*.py",
    "README.md",
    "*.md"
]

[tool.hatch.build.targets.sdist]
include = [
    "app/",
    "*.py",
    "*.md",
    "pyproject.toml"
]

[tool.uv]
dev-dependencies = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--verbose",
    "--tb=short",
    "--disable-warnings",
]
